+++ framework_src/Attribute/TestSuite.php
@@ -4,6 +4,11 @@ namespace Cspray\Labrador\AsyncUnit\Attribute;
 
 use Attribute;
 
+/**
+ * Class TestSuite
+ * @package Cspray\Labrador\AsyncUnit\Attribute
+ * @codeCoverageIgnore
+ */
 #[Attribute(Attribute::TARGET_CLASS)]
 final class TestSuite {