+++ framework_src/Exception/TestDisabledException.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Cspray\Labrador\AsyncUnit\Exception;
+
+
+use Throwable;
+
+final class TestDisabledException extends Exception {
+
+    public function __construct($message) {
+        parent::__construct($message);
+    }
+
+}