+++ framework_src/TestFrameworkApplication.php
@@ -12,7 +12,7 @@ use Cspray\Labrador\AsyncUnit\Event\TestProcessingStartedEvent;
 use Cspray\Labrador\AsyncUnit\Exception\AssertionFailedException;
 use Cspray\Labrador\AsyncUnit\Exception\InvalidStateException;
 use Cspray\Labrador\AsyncUnit\Exception\TestFailedException;
-use Cspray\Labrador\AsyncUnit\Event\TestInvokedEvent;
+use Cspray\Labrador\AsyncUnit\Event\TestProcessedEvent;
 use Cspray\Labrador\Plugin\Pluggable;
 use stdClass;
 use function Amp\call;
@@ -44,7 +44,7 @@ final class TestFrameworkApplication extends AbstractApplication {
             $testRunState->totalAssertionCount = 0;
             $testRunState->totalAsyncAssertionCount = 0;
 
-            $this->emitter->on(Events::TEST_INVOKED, function(TestInvokedEvent $testInvokedEvent) use($testRunState) {
+            $this->emitter->on(Events::TEST_PROCESSED, function(TestProcessedEvent $testInvokedEvent) use($testRunState) {
                 $testRunState->testsInvoked++;
                 $testRunState->totalAssertionCount += $testInvokedEvent->getTarget()->getTestCase()->getAssertionCount();
                 $testRunState->totalAsyncAssertionCount += $testInvokedEvent->getTarget()->getTestCase()->getAsyncAssertionCount();
@@ -103,40 +103,6 @@ final class TestFrameworkApplication extends AbstractApplication {
 
             public function getAsyncAssertionCount() : int {
                 return $this->testRunState->totalAsyncAssertionCount;
-            }
-        };
-    }
-
-    private function getTestResult(
-        TestCase $testCase,
-        string $method,
-        ?TestFailedException $testFailedException
-    ) : TestResult {
-        return new class($testCase, $method, $testFailedException) implements TestResult {
-
-            public function __construct(
-                private TestCase $testCase,
-                private string $method,
-                private ?TestFailedException $testFailedException
-            ) {}
-
-            public function getTestCase() : TestCase {
-                return $this->testCase;
-            }
-
-            public function getTestMethod() : string {
-                return $this->method;
-            }
-
-            public function isSuccessful() : bool {
-                return is_null($this->testFailedException);
-            }
-
-            public function getFailureException() : TestFailedException|AssertionFailedException {
-                if (is_null($this->testFailedException)) {
-                    throw new InvalidStateException('Attempted to access a TestFailedException on a successful TestResult.');
-                }
-                return $this->testFailedException;
             }
         };
     }